+++ kernel/src/utils/memory.cpp
@@ -6,7 +6,7 @@ uint64_t GetMemorySize(EFI_MEMORY_DESCRIPTOR* mMap, uint64_t mMapEntries, uint64
     static uint64_t memorySizeBytes = 0;
     if (memorySizeBytes > 0) return memorySizeBytes;
 
-    for (int i = 0; i < mMapEntries; i++){
+    for (uint64_t i = 0; i < mMapEntries; i++){
         EFI_MEMORY_DESCRIPTOR* desc = (EFI_MEMORY_DESCRIPTOR*)((uint64_t)mMap + (i * mMapDescSize));
         memorySizeBytes += desc->numPages * 4096;
     }
